better example classes for specs
As we added more examples and permutations to the specs, the classes
that were there didn't make sense any more. (Why would gentlemen and 
gentlewomen have gloves, but not just men and women?).

The new classes are easier to understand, more fun to work with, and 
are more easily expanded upon as we want to test more situations.