specify db table on has_many sql conditions
`has_many_as_polymorphs` uses a Proc to set conditions on a has_many 
relationship. On Rails 3.2, when this is used in conjunction with a
has_many :through relationship, the generated SQL specifies the wrong
table name for the conditions generated by `has_many_as_polymorphs`.

This commit fixes the issue by making `has_many_as_polymorphs` generate
conditions that specify the table they should be applied to.