avoid exception with Postgres
Postgres is still not supported, but at the very least we should
make sure the gem doesn’t cause exceptions to be raised when used
alongside Postgres. This is relevant for a situation where we have
Polymorpheus in our Gemfile because our primary db is MySQL, but 
we are also using a secondary Postgres db.