Remove duplicate statements from schema
This commit fixes a bug in SchemaDumper that was causing all
`add_polymorphic_triggers` statements to be duplicated. Since this
method adds triggers for both CREATE and UPDATE, two triggers are
created each time. When dumping the schema, we should de-duplicate these
statements.

The specs for this class have also been updated to use real database
queries instead of stubbing the `connection.triggers` method.