Remove support for ActiveRecord beyond EOL
This commit removes all appraisals for unsupported versions of
ActiveRecord. Previous versions of this gem should still work with those
versions, but we will not attempt to support every published version of
ActiveRecord going forward.

By removing support for these very old versions, we are also able to
remove many `ActiveRecord::VERSION` checks and code that was supporting
ActiveRecord <= 5.2.