Add support for ActiveRecord 6.1
This commit adds an appraisal for activerecord 6.1 and increases the
maximum version allowed by polymorpheus.

In order to support changes to the AR public API, we will also switch to
using `ActiveRecord::Base.connection_db_config` to retrieve the
`db_config` object from the current database connection when we are
using a version of ActiveRecord that supports this interface.